item list crud, optimization, keyboard handling, query updates
item list crud, optimization, keyboard handling, query updates

+++ app/domain/stores/noteStore.ts
@@ -6,6 +6,7 @@ import { DIName, RECORD } from '../constants'
 import { unixDateConverter } from '../helper'
 import { NoteType } from '../enums'
 import NoteItemModel from '../models/noteItem'
+import { Item } from 'react-native-paper/lib/typescript/src/components/List/List'
 
 const defaults = {
   loading: false,
@@ -42,35 +43,73 @@ export class NoteStore {
   }
 
   @action.bound addListItem(noteItemModel: NoteItemModel) {
-    noteItemModel.noteId = this.header.noteModel.id
-    this.header.noteModel.items!.push(noteItemModel)
+    let nextIndex = 0
+    if (this.header.noteModel.items!.length > 0) {
+      const indices = this.header.noteModel.items!.map((item: NoteItemModel) => item.rowIndex)
+      nextIndex = Math.max(...indices) + 1
+    }
 
-    console.log(this.header.noteModel.items)
+    noteItemModel.rowIndex = nextIndex
+    noteItemModel.noteId = this.header.noteModel.id
+    this.unselectItems()
+    noteItemModel.selected = true
+
+    this.header.noteModel.items!.push(noteItemModel)
+  }
+
+  @action.bound selectItem(noteItemModel: NoteItemModel) {
+    this.unselectItems()
+    const target = this.header.noteModel.items!.find(item => item.rowIndex == noteItemModel.rowIndex)
+    if (target) {
+      target.selected = true
+    }
+  }
+
+  @action.bound unselectItems() {
+    for (let index = 0; index < this.header.noteModel.items!.length; index++) {
+      this.header.noteModel.items![index].selected = false
+    }
+  }
+
+  @action.bound removeSelectedItem() {
+    const notSeletedItems = this.header.noteModel.items?.filter(item => !item.selected)
+    if (notSeletedItems) {
+      this.header.noteModel.items = notSeletedItems
+    }
+  }
+
+  @action.bound checkedItem(noteItemModel: NoteItemModel, value: boolean) {
+    this.header.noteModel.items?.slice()
+    const target = this.header.noteModel.items!.find(item => item.rowIndex == noteItemModel.rowIndex)
+    if (target) {
+      target.checked = value
+      this.refreshListItem()
+    }
+  }
+
+  @action.bound refreshListItem() {
+    this.header.noteModel.items = this.header.noteModel.items!.filter(_obj => true)
   }
 
   @action.bound async edit(model: NoteModel) {
     this.header.isEditing = true
     this.header.isNew = false
 
-    this.header.oldNoteModel = <NoteModel>{ ...model }
-    this.header.noteModel = model
-
     if (model.type == NoteType.List) {
       var result = await this.noteItemService.getAll({
-        order: 'index DESC',
+        order: 'rowIndex DESC',
         where: {
           noteId_eq: model.id,
         },
       })
 
-      if (result.data.length > 0) {
-        this.header.noteModel.items = observable.array(
-          (result.data as []).map((data: any) => <NoteItemModel>{ ...data })
-        )
-      } else {
-        this.header.noteModel.items = observable.array(<NoteItemModel[]>[])
+      if (result.successful) {
+        model.items = result.data.map((dataModel: any) => new NoteItemModel({ ...dataModel }))
       }
     }
+
+    this.header.oldNoteModel = <NoteModel>{ ...model }
+    this.header.noteModel = model
   }
 
   @action.bound add(model: NoteModel) {
@@ -78,7 +117,7 @@ export class NoteStore {
     this.header.isNew = true
 
     if (model.type == NoteType.List) {
-      model.items = observable.array(<NoteItemModel[]>[])
+      model.items = <NoteItemModel[]>[]
     }
 
     this.header.noteModel = model
@@ -95,13 +134,34 @@ export class NoteStore {
       limit: RECORD.defaultLimit,
       page: RECORD.defaultPage,
     })
-    this.header.loading = false
     // console.log(all)
     if (all && all.data.length > 0) {
-      this.header.notes = all.data.map((model: any) => new NoteModel({ ...model }))
+      let tempNotes = all.data.map((noteModel: any) => new NoteModel({ ...noteModel }))
+
+      for (let index = 0; index < tempNotes.length; index++) {
+        let note = tempNotes[index]
+        if (note.type == NoteType.List) {
+          var result = await this.noteItemService.getAll({
+            limit: 3,
+            page: 1,
+            order: 'rowIndex DESC',
+            where: {
+              noteId_eq: note.id,
+            },
+          })
+
+          if (result.successful) {
+            note.items = result.data.map((noteItem: any) => new NoteItemModel({ ...noteItem }))
+          }
+        }
+      }
+
+      this.header.notes = tempNotes
     } else {
       this.header.notes = Array<NoteModel>()
     }
+
+    this.header.loading = false
   }
 
   @action.bound async delete(noteModel: NoteModel) {
@@ -110,15 +170,31 @@ export class NoteStore {
     }
 
     this.header.loading = true
-    var result = await this.noteService.delete(<NoteModel>{ ...noteModel })
-    this.header.loading = false
-    this.header.isEditing = false
-    this.header.isNew = false
-    this.header.noteModel = <NoteModel>{}
-    this.header.showAddOption = true
 
-    if (result.successful) {
-      return true
+    const resultItems = await this.noteItemService.getAll({
+      where: {
+        noteId_eq: noteModel.id,
+      },
+    })
+
+    if (resultItems.successful) {
+      //delete if it has items
+      if (resultItems && resultItems.data.length > 0) {
+        resultItems.data.forEach(async (item: NoteItemModel) => await this.noteItemService.delete(item))
+      }
+
+      var result = await this.noteService.delete(<NoteModel>{ ...noteModel })
+      if (result.successful) {
+        this.header.loading = false
+        this.header.isEditing = false
+        this.header.isNew = false
+        this.header.noteModel = <NoteModel>{}
+        this.header.showAddOption = true
+
+        if (result.successful) {
+          return true
+        }
+      }
     }
 
     return false
@@ -131,25 +207,38 @@ export class NoteStore {
 
     this.header.loading = true
 
+    //make sure one of the field has value
+    if (!this.header.noteModel.title) {
+      this.header.noteModel.title = unixDateConverter(Date.now())
+    } else {
+      this.header.noteModel.title = this.header.noteModel.title.trim()
+    }
+
     //New record
-    if (this.header.noteModel.id === null || this.header.noteModel.id === 0 || this.header.noteModel.id === undefined) {
+    const isNewRecord =
+      this.header.noteModel.id === null || this.header.noteModel.id === 0 || this.header.noteModel.id === undefined
+    if (isNewRecord) {
       this.header.noteModel.createDate = Date.now()
       this.header.noteModel.modifiedDate = Date.now()
 
-      //make sure one of the field has value
-      if (this.header.noteModel.title || this.header.noteModel.content) {
-        if (!this.header.noteModel.title) {
-          this.header.noteModel.title = unixDateConverter(this.header.noteModel.createDate)
-        } else {
-          this.header.noteModel.title = this.header.noteModel.title.trim()
-        }
+      var result = await this.noteService.add(<NoteModel>{ ...this.header.noteModel })
+      if (result.successful) {
+        console.log('New Note successful', result)
 
-        var result = await this.noteService.add(<NoteModel>{ ...this.header.noteModel })
-        if (!result.successful) {
-          console.log('New save failed', result)
-        }
+        if (this.header.noteModel.type == NoteType.List) {
+          this.header.noteModel.items?.forEach(async item => {
+            await this.noteItemService.add(<NoteItemModel>{
+              noteId: result.data.id,
+              rowIndex: item.rowIndex,
+              checked: item.checked,
+              title: item.title,
+              createDate: Date.now(),
+              modifiedDate: Date.now(),
+            })
+          })
 
-        console.log('New save successful')
+          console.log('New items successful')
+        }
       }
 
       //Update record
@@ -157,23 +246,46 @@ export class NoteStore {
       if (
         this.header.noteModel.title === this.header.oldNoteModel.title &&
         this.header.noteModel.content === this.header.oldNoteModel.content &&
-        this.header.noteModel.rank === this.header.oldNoteModel.rank
+        this.header.noteModel.rank === this.header.oldNoteModel.rank &&
+        this.header.noteModel.type == NoteType.Note
       ) {
         this.header.isEditing = false
         this.header.loading = false
         return
       }
 
-      if (!this.header.noteModel.title) {
-        this.header.noteModel.title = unixDateConverter(Date.now())
-      } else {
-        this.header.noteModel.title = this.header.noteModel.title.trim()
-      }
-
       this.header.noteModel.modifiedDate = Date.now()
       var result = await this.noteService.update(<NoteModel>{ ...this.header.noteModel })
-      if (!result.successful) {
-        console.log('New update failed', result)
+      if (result.successful) {
+        console.log(this.header.noteModel.id)
+
+        if (this.header.noteModel.type == NoteType.List) {
+          //delete the items already exists
+          const result = await this.noteItemService.getAll({
+            where: {
+              noteId_eq: this.header.noteModel.id,
+            },
+          })
+
+          if (result.successful) {
+            if (result.data && result.data.length > 0) {
+              result.data.forEach(async (item: NoteItemModel) => {
+                await this.noteItemService.delete(item)
+              })
+            }
+
+            this.header.noteModel.items?.forEach(async item => {
+              await this.noteItemService.add(<NoteItemModel>{
+                noteId: this.header.noteModel.id,
+                rowIndex: item.rowIndex,
+                checked: item.checked,
+                title: item.title,
+                createDate: Date.now(),
+                modifiedDate: Date.now(),
+              })
+            })
+          }
+        }
       }
 
       console.log('New update successful')
